## Summary
- Add opt-in scan-ahead conflict detection for Copy/Move operations with
dependency-aware ready and pending queues.
- Preserve safe independent transfers and parallel copy progress while
conflicts await user decisions.
- Add responsive conflict operations UI, operation-wide bulk actions,
stable progress layout, and automatic next-row selection.
- Move Copy/Move preferences to the File Operations configuration page
and add focused contract/state tests.

## Test plan
- [x] Focused conflict-state tests
- [x] Progress-layout tests
- [x] Scheduling contract tests
- [x] SFX build-race contract tests
- [x] x64 Debug ClCompile and Link
- [x] git diff --check
- [x] Manual runtime verification of scan-ahead Copy/Move flows

+++ src/plugins/zip/vcxproj/selfextr/makeall.bat
@@ -1,5 +1,10 @@
 @echo off
 
+setlocal
+set "build_exit_code=1"
+set "sfx_build_lock_acquired="
+set "sfx_pushd_done="
+
 REM
 REM Usage: makeall.bat
 REM
@@ -8,28 +13,28 @@ REM Build the SFX package into the sfx subdirectory
 if "%OPENSAL_BUILD_DIR%"=="" (
   echo Please set OPENSAL_BUILD_DIR environment variable.
   echo.
-  pause
-  exit /b
+  goto cleanup
 )
 
 if not exist "%OPENSAL_BUILD_DIR%salamander" (
   echo Build directory does not exist: %OPENSAL_BUILD_DIR%salamander
   echo.
-  pause
-  exit /b
+  goto cleanup
 )
 
 set "sfx_build_lock=%TEMP%\OpenSalamander-SFX-build.lock"
 :acquire_sfx_build_lock
 2>nul mkdir "%sfx_build_lock%"
-if not exist "%sfx_build_lock%" (
+if errorlevel 1 (
   echo Another SFX build is using the shared self-extractor outputs. Waiting...
   CHOICE /T 1 /C yn /D y >nul
   goto acquire_sfx_build_lock
 )
+set "sfx_build_lock_acquired=yes"
 
-setlocal
 pushd "%~dp0"
+if errorlevel 1 goto build_failed
+set "sfx_pushd_done=yes"
 set making_all_versions=dummy
 
 echo Making all language versions.
@@ -39,6 +44,7 @@ for %%v in (ENGLISH_VERSION CZECH_VERSION SLOVAK_VERSION GERMAN_VERSION ^
             RUSSIAN_VERSION CHINESESIMPL_VERSION DUTCH_VERSION ^
             FRENCH_VERSION ITALIAN_VERSION) do (
   call "%~dp0sfxmake.bat" %%v sfx\
+  if errorlevel 1 goto build_failed
 )
 
 echo.
@@ -61,57 +67,67 @@ if "%makeall_should_pause%"=="" (
 for %%t in (Debug_x86 Release_x86 Debug_x64 Release_x64 Release_arm64 "Release clean_x64" "Release clean_arm64") do (
   echo.
   echo Copying SFX packages to %%~t...
-  if not exist "%OPENSAL_BUILD_DIR%salamander\%%~t\plugins\zip\sfx" mkdir "%OPENSAL_BUILD_DIR%salamander\%%~t\plugins\zip\sfx"
+  if not exist "%OPENSAL_BUILD_DIR%salamander\%%~t\plugins\zip\sfx" (
+    mkdir "%OPENSAL_BUILD_DIR%salamander\%%~t\plugins\zip\sfx"
+    if errorlevel 1 goto build_failed
+  )
   for %%s in (czech.sfx english.sfx german.sfx slovak.sfx spanish.sfx ^
               romanian.sfx hungarian.sfx russian.sfx chinesesimplified.sfx ^
               dutch.sfx french.sfx italian.sfx) do (
-    @call :sfxcopy "sfx\%%s" "%OPENSAL_BUILD_DIR%salamander\%%~t\plugins\zip\sfx\%%s"
+    call :sfxcopy "sfx\%%s" "%OPENSAL_BUILD_DIR%salamander\%%~t\plugins\zip\sfx\%%s"
+    if errorlevel 1 goto build_failed
   )
 )
 
 if exist sfx rmdir /Q /S sfx
+set "build_exit_code=0"
+goto cleanup
 
-popd
-endlocal
-rmdir /Q /S "%sfx_build_lock%" >nul 2>nul
+:build_failed
+set "build_exit_code=%errorlevel%"
+if "%build_exit_code%"=="0" set "build_exit_code=1"
+echo.
+echo SFX build failed with exit code %build_exit_code%.
+
+:cleanup
+if "%sfx_pushd_done%"=="yes" popd
+if "%sfx_build_lock_acquired%"=="yes" rmdir /Q /S "%sfx_build_lock%" >nul 2>nul
 if "%makeall_should_pause%"=="" (
   echo.
   pause
 )
-@exit /b
+endlocal & exit /b %build_exit_code%
 
 
-rem ---------------------------- Copy routine
+rem ----------------------------- Copy routine
 
 :sfxcopy
 
 copy /Y %1 %2
-if not errorlevel 1 goto END_copy
-echo !!!!!!!!!!!!!!!!!!!   ERROR   !!!!!!!!!!!!!!!!!!!
-echo !!!!!!!!!!!!!!!!!!!   ERROR   !!!!!!!!!!!!!!!!!!!
-echo !!!!!!!!!!!!!!!!!!!   ERROR   !!!!!!!!!!!!!!!!!!!
-pause
+if errorlevel 1 (
+  echo !!!!!!!!!!!!!!!!!!!!   ERROR   !!!!!!!!!!!!!!!!!!!!
+  echo !!!!!!!!!!!!!!!!!!!!   ERROR   !!!!!!!!!!!!!!!!!!!!
+  echo !!!!!!!!!!!!!!!!!!!!   ERROR   !!!!!!!!!!!!!!!!!!!!
+  if "%makeall_should_pause%"=="" pause
+  exit /b 1
+)
+exit /b 0
 
-:END_copy
-
-exit /b
-
-rem ---------------------------- Move routine
+rem ----------------------------- Move routine
 
 :sfxmove
 
 move /Y %1 %2
-if not errorlevel 1 goto END_move
-echo !!!!!!!!!!!!!!!!!!!   ERROR   !!!!!!!!!!!!!!!!!!!
-echo !!!!!!!!!!!!!!!!!!!   ERROR   !!!!!!!!!!!!!!!!!!!
-echo !!!!!!!!!!!!!!!!!!!   ERROR   !!!!!!!!!!!!!!!!!!!
-pause
+if errorlevel 1 (
+  echo !!!!!!!!!!!!!!!!!!!!   ERROR   !!!!!!!!!!!!!!!!!!!!
+  echo !!!!!!!!!!!!!!!!!!!!   ERROR   !!!!!!!!!!!!!!!!!!!!
+  echo !!!!!!!!!!!!!!!!!!!!   ERROR   !!!!!!!!!!!!!!!!!!!!
+  if "%makeall_should_pause%"=="" pause
+  exit /b 1
+)
+exit /b 0
 
-:END_move
-
-exit /b
-
-rem ---------------------------- Remove directory with auto-retry routine
+rem ----------------------------- Remove directory with auto-retry routine
 
 :remove_dir_autoretry
 if exist %1 rmdir /Q /S %1
